+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 # 2.2.2
 * bugfix KeywordExtractor: Also put Character Class M,Cs,Co to L
 * bugfix KeywordExtractor: works with upper, part and alpha
-
+* update KwsError: delete unused options
 ## 2.2.1
 * bugfix Keyword extractor
 * improvements of tests